+++ waspc/data/Generator/templates/react-app/src/auth/email/index.ts
@@ -1,0 +1,8 @@
+export { login } from './actions/login';
+export { signup } from './actions/signup';
+export { requestPasswordReset } from './actions/passwordReset';
+export { resetPassword } from './actions/passwordReset';
+export { verifyEmail } from './actions/verifyEmail';
+
+export { SignupForm } from './components/Signup';
+export { LoginForm } from './components/Login';