+++ waspc/data/Generator/templates/react-app/src/auth/forms/Login.jsx
@@ -1,46 +1,14 @@
 {{={= =}=}}
-import React, { useState } from 'react'
-import { useHistory } from 'react-router-dom'
+import Auth from './Auth'
 
-import login from '../login.js'
-import { errorMessage } from '../../utils.js'
-
-const LoginForm = () => {
-  const history = useHistory()
-
-  const [usernameFieldVal, setUsernameFieldVal] = useState('')
-  const [passwordFieldVal, setPasswordFieldVal] = useState('')
-
-  const handleLogin = async (event) => {
-    event.preventDefault()
-    try {
-      await login(usernameFieldVal, passwordFieldVal)
-      // Redirect to configured page, defaults to /.
-      history.push('{= onAuthSucceededRedirectTo =}')
-    } catch (err) {
-      console.log(err)
-      window.alert(errorMessage(err))
-    }
-  }
-  
+const LoginForm = ({ appearance, logo, socialLayout }) => {
   return (
-    <form onSubmit={handleLogin} className="login-form auth-form">
-      <h2>Username</h2>
-      <input
-        type="text"
-        value={usernameFieldVal}
-        onChange={e => setUsernameFieldVal(e.target.value)}
-      />
-      <h2>Password</h2>
-      <input
-        type="password"
-        value={passwordFieldVal}
-        onChange={e => setPasswordFieldVal(e.target.value)}
-      />
-      <div>
-        <input type="submit" value="Log in"/>
-      </div>
-    </form>
+    <Auth
+      appearance={appearance}
+      logo={logo}
+      socialLayout={socialLayout}
+      isLogin={true}
+    />
   )
 }