+++ waspc/data/Generator/templates/server/src/auth/providers/email/types.ts
@@ -1,0 +1,13 @@
+export type GetVerificationEmailContentFn = (params: { verificationLink: string }) => EmailContent;
+
+export type GetPasswordResetEmailContentFn = (params: { passwordResetLink: string }) => EmailContent;
+
+type EmailContent = {
+    subject: string;
+    html: string;
+    text: string;
+}
+
+export const tokenVerificationErrors = {
+    TokenExpiredError: 'TokenExpiredError',
+};