+++ waspc/e2e-test/test-outputs/waspComplexTest-golden/waspComplexTest/.wasp/out/server/src/dbClient.ts
@@ -1,15 +1,11 @@
 import Prisma from '@prisma/client'
 
-
 import { registerAuthMiddleware } from './core/auth/prismaMiddleware.js'
-
 
 const createDbClient = () => {
   const prismaClient = new Prisma.PrismaClient()
 
-
   registerAuthMiddleware(prismaClient)
-
 
   return prismaClient
 }