+++ waspc/src/Wasp/AppSpec/App/Auth.hs
@@ -5,16 +5,22 @@ module Wasp.AppSpec.App.Auth
   ( Auth (..),
     AuthMethods (..),
     ExternalAuthConfig (..),
+    EmailAuthConfig (..),
     usernameAndPasswordConfig,
     isUsernameAndPasswordAuthEnabled,
+    areBothExternalAndUsernameAndPasswordAuthEnabled,
     isExternalAuthEnabled,
     isGoogleAuthEnabled,
     isGitHubAuthEnabled,
+    isEmailAuthEnabled,
   )
 where
 
 import Data.Data (Data)
 import Data.Maybe (isJust)
+import Wasp.AppSpec.App.Auth.EmailVerification (EmailVerificationConfig)
+import Wasp.AppSpec.App.Auth.PasswordReset (PasswordResetConfig)
+import Wasp.AppSpec.App.EmailSender (EmailFromField)
 import Wasp.AppSpec.Core.Ref (Ref)
 import Wasp.AppSpec.Entity (Entity)
 import Wasp.AppSpec.ExtImport (ExtImport)
@@ -31,7 +37,8 @@ data Auth = Auth
 data AuthMethods = AuthMethods
   { usernameAndPassword :: Maybe UsernameAndPasswordConfig,
     google :: Maybe ExternalAuthConfig,
-    gitHub :: Maybe ExternalAuthConfig
+    gitHub :: Maybe ExternalAuthConfig,
+    email :: Maybe EmailAuthConfig
   }
   deriving (Show, Eq, Data)
 
@@ -47,11 +54,22 @@ data ExternalAuthConfig = ExternalAuthConfig
   }
   deriving (Show, Eq, Data)
 
+data EmailAuthConfig = EmailAuthConfig
+  { fromField :: EmailFromField,
+    emailVerification :: EmailVerificationConfig,
+    passwordReset :: PasswordResetConfig,
+    allowUnverifiedLogin :: Maybe Bool
+  }
+  deriving (Show, Eq, Data)
+
 usernameAndPasswordConfig :: UsernameAndPasswordConfig
 usernameAndPasswordConfig = UsernameAndPasswordConfig Nothing
 
 isUsernameAndPasswordAuthEnabled :: Auth -> Bool
 isUsernameAndPasswordAuthEnabled = isJust . usernameAndPassword . methods
+
+areBothExternalAndUsernameAndPasswordAuthEnabled :: Auth -> Bool
+areBothExternalAndUsernameAndPasswordAuthEnabled auth = all ($ auth) [isExternalAuthEnabled, isUsernameAndPasswordAuthEnabled]
 
 isExternalAuthEnabled :: Auth -> Bool
 isExternalAuthEnabled auth = any ($ auth) [isGoogleAuthEnabled, isGitHubAuthEnabled]
@@ -61,3 +79,6 @@ isGoogleAuthEnabled = isJust . google . methods
 
 isGitHubAuthEnabled :: Auth -> Bool
 isGitHubAuthEnabled = isJust . gitHub . methods
+
+isEmailAuthEnabled :: Auth -> Bool
+isEmailAuthEnabled = isJust . email . methods