+++ waspc/src/Wasp/Generator/ServerGenerator.hs
@@ -13,10 +13,12 @@ where
 
 import Data.Aeson (object, (.=))
 import qualified Data.Aeson as Aeson
+import qualified Data.ByteString.Lazy.UTF8 as ByteStringLazyUTF8
 import Data.Maybe
   ( fromJust,
     fromMaybe,
     isJust,
+    maybeToList,
   )
 import StrongPath
   ( Dir,
@@ -56,6 +58,7 @@ import Wasp.Generator.ServerGenerator.Auth.OAuthAuthG (depsRequiredByPassport)
 import Wasp.Generator.ServerGenerator.AuthG (genAuth)
 import qualified Wasp.Generator.ServerGenerator.Common as C
 import Wasp.Generator.ServerGenerator.ConfigG (genConfigFile)
+import Wasp.Generator.ServerGenerator.Db.Seed (genDbSeed, getPackageJsonPrismaSeedField)
 import Wasp.Generator.ServerGenerator.EmailSenderG (depsRequiredByEmail, genEmailSender)
 import Wasp.Generator.ServerGenerator.ExternalCodeGenerator (extServerCodeGeneratorStrategy, extSharedCodeGeneratorStrategy)
 import Wasp.Generator.ServerGenerator.JobGenerator (depsRequiredByJobs, genJobExecutors, genJobs)
@@ -127,11 +130,17 @@ genPackageJson spec waspDependencies = do
                 .= ( (if hasEntities then "npm run db-migrate-prod && " else "")
                        ++ "NODE_ENV=production npm run start"
                    ),
-              "overrides" .= getPackageJsonOverrides
+              "overrides" .= getPackageJsonOverrides,
+              "prisma" .= ByteStringLazyUTF8.toString (Aeson.encode $ getPackageJsonPrismaField spec)
             ]
       )
   where
     hasEntities = not . null $ AS.getDecls @AS.Entity.Entity spec
+
+getPackageJsonPrismaField :: AppSpec -> Aeson.Value
+getPackageJsonPrismaField spec = object $ [] <> seedEntry
+  where
+    seedEntry = maybeToList $ Just . ("seed" .=) =<< getPackageJsonPrismaSeedField spec
 
 npmDepsForWasp :: AppSpec -> N.NpmDepsForWasp
 npmDepsForWasp spec =
@@ -151,6 +160,7 @@ npmDepsForWasp spec =
             ("patch-package", "^6.4.7"),
             ("uuid", "^9.0.0"),
             ("lodash.merge", "^4.6.2"),
+            ("rate-limiter-flexible", "^2.4.1"),
             ("superjson", "^1.12.2")
           ]
           ++ depsRequiredByPassport spec
@@ -204,6 +214,7 @@ genSrcDir spec =
     <++> genOperations spec
     <++> genAuth spec
     <++> genEmailSender spec
+    <++> genDbSeed spec
   where
     genFileCopy = return . C.mkSrcTmplFd
 
@@ -212,7 +223,7 @@ genDbClient spec = return $ C.mkTmplFdWithDstAndData tmplFile dstFile (Just tmpl
   where
     maybeAuth = AS.App.auth $ snd $ getApp spec
 
-    dbClientRelToSrcP = [relfile|dbClient.js|]
+    dbClientRelToSrcP = [relfile|dbClient.ts|]
     tmplFile = C.asTmplFile $ [reldir|src|] </> dbClientRelToSrcP
     dstFile = C.serverSrcDirInServerRootDir </> C.asServerSrcFile dbClientRelToSrcP
 
@@ -361,6 +372,7 @@ genExportedTypesDir spec =
     [ C.mkTmplFdWithData [relfile|src/types/index.ts|] (Just tmplData)
     ]
   where
-    tmplData = object ["isExternalAuthEnabled" .= isExternalAuthEnabled]
+    tmplData = object ["isExternalAuthEnabled" .= isExternalAuthEnabled, "isEmailAuthEnabled" .= isEmailAuthEnabled]
     isExternalAuthEnabled = AS.App.Auth.isExternalAuthEnabled <$> maybeAuth
+    isEmailAuthEnabled = AS.App.Auth.isEmailAuthEnabled <$> maybeAuth
     maybeAuth = AS.App.auth $ snd $ getApp spec