+++ waspc/src/Wasp/Generator/WebAppGenerator/AuthG.hs
@@ -4,13 +4,18 @@ module Wasp.Generator.WebAppGenerator.AuthG
 where
 
 import Data.Aeson (object, (.=))
-import StrongPath (relfile)
+import Data.Aeson.Types (Pair)
+import StrongPath (File', Path', Rel', reldir, relfile, (</>))
 import Wasp.AppSpec (AppSpec)
 import qualified Wasp.AppSpec.App as AS.App
 import qualified Wasp.AppSpec.App.Auth as AS.Auth
 import Wasp.AppSpec.Valid (getApp)
+import Wasp.Generator.AuthProviders (gitHubAuthProvider, googleAuthProvider)
+import qualified Wasp.Generator.AuthProviders.OAuth as OAuth
 import Wasp.Generator.FileDraft (FileDraft)
 import Wasp.Generator.Monad (Generator)
+import Wasp.Generator.WebAppGenerator.Auth.Common (getOnAuthSucceededRedirectToOrDefault)
+import Wasp.Generator.WebAppGenerator.Auth.EmailAuthG (genEmailAuth)
 import Wasp.Generator.WebAppGenerator.Auth.LocalAuthG (genLocalAuth)
 import Wasp.Generator.WebAppGenerator.Auth.OAuthAuthG (genOAuthAuth)
 import Wasp.Generator.WebAppGenerator.Common as C
@@ -26,8 +31,10 @@ genAuth spec =
           genCreateAuthRequiredPage auth,
           genUserHelpers
         ]
+        <++> genAuthForms auth
         <++> genLocalAuth auth
         <++> genOAuthAuth auth
+        <++> genEmailAuth auth
     Nothing -> return []
   where
     maybeAuth = AS.App.auth $ snd $ getApp spec
@@ -49,6 +56,55 @@ genCreateAuthRequiredPage auth =
 --   ot not).
 genUseAuth :: Generator FileDraft
 genUseAuth = return $ C.mkTmplFd (C.asTmplFile [relfile|src/auth/useAuth.ts|])
+
+genAuthForms :: AS.Auth.Auth -> Generator [FileDraft]
+genAuthForms auth =
+  sequence
+    [ genLoginForm auth,
+      genSignupForm auth,
+      genAuthForm auth,
+      copyTmplFile [relfile|stitches.config.js|],
+      copyTmplFile [relfile|auth/forms/SocialIcons.tsx|]
+    ]
+  where
+    copyTmplFile = return . C.mkSrcTmplFd
+
+genAuthForm :: AS.Auth.Auth -> Generator FileDraft
+genAuthForm auth =
+  compileTmplToSamePath
+    [relfile|auth/forms/Auth.tsx|]
+    [ "onAuthSucceededRedirectTo" .= getOnAuthSucceededRedirectToOrDefault auth,
+      "isUsernameAndPasswordAuthEnabled" .= AS.Auth.isUsernameAndPasswordAuthEnabled auth,
+      "areBothExternalAndUsernameAndPasswordAuthEnabled" .= AS.Auth.areBothExternalAndUsernameAndPasswordAuthEnabled auth,
+      "isExternalAuthEnabled" .= AS.Auth.isExternalAuthEnabled auth,
+      -- Google
+      "isGoogleAuthEnabled" .= AS.Auth.isGoogleAuthEnabled auth,
+      "googleSignInPath" .= OAuth.serverLoginUrl googleAuthProvider,
+      -- GitHub
+      "isGitHubAuthEnabled" .= AS.Auth.isGitHubAuthEnabled auth,
+      "gitHubSignInPath" .= OAuth.serverLoginUrl gitHubAuthProvider
+    ]
+
+genLoginForm :: AS.Auth.Auth -> Generator FileDraft
+genLoginForm auth =
+  compileTmplToSamePath
+    [relfile|auth/forms/Login.jsx|]
+    [ "onAuthSucceededRedirectTo" .= getOnAuthSucceededRedirectToOrDefault auth
+    ]
+
+genSignupForm :: AS.Auth.Auth -> Generator FileDraft
+genSignupForm auth =
+  compileTmplToSamePath
+    [relfile|auth/forms/Signup.jsx|]
+    [ "onAuthSucceededRedirectTo" .= getOnAuthSucceededRedirectToOrDefault auth
+    ]
+
+compileTmplToSamePath :: Path' Rel' File' -> [Pair] -> Generator FileDraft
+compileTmplToSamePath tmplFileInTmplSrcDir keyValuePairs =
+  return $
+    C.mkTmplFdWithData
+      (asTmplFile $ [reldir|src|] </> tmplFileInTmplSrcDir)
+      (object keyValuePairs)
 
 genUserHelpers :: Generator FileDraft
 genUserHelpers = return $ C.mkTmplFd (C.asTmplFile [relfile|src/auth/helpers/user.ts|])